webpack bundling and new sezzle.js file added
