+++ gulpfile.js
@@ -8,27 +8,18 @@ var gulp = require('gulp'),
     rimraf = require('gulp-rimraf'),
     s3 = require('gulp-s3-upload')(config),
     cloudfront = require('gulp-cloudfront-invalidate'),
-    rp = require("request-promise"),
-    config = { useIAM: true };
+    rp = require('request-promise'),
+    config = { useIAM: true },
+    webpack = require('webpack-stream'),
+    pjson = require('./package.json');
 
-var buttonUploadName = 'sezzle-widget2.3.3.js';
-var bannerUploadName = 'sezzle-banner2.1.1.js';
+var buttonUploadName = `sezzle-widget${pjson.version}.js`;
 var globalCssUploadName = 'sezzle-styles-global2.0.2.css';
 
-gulp.task("cssupload", function() {
-    // bucket base url https://d3svog4tlx445w.cloudfront.net/
-    var indexPath = './dist/global-css/global.min.css'
-    gulp.src(indexPath)
-        .pipe(rename('shopify-app/assets/' + globalCssUploadName))
-        .pipe(s3({
-            Bucket: 'sezzlemedia', //  Required
-            ACL: 'public-read'       //  Needs to be user-defined
-        }, {
-            maxRetries: 5
-        }))
-});
-
-gulp.task("upload-initial", function() {
+/**
+ * Tasks for the initial script
+ */
+gulp.task('upload-initial', function() {
     var indexPath = './initial-script.js';
     gulp.src(indexPath)
         .pipe(s3({
@@ -47,56 +38,13 @@ gulp.task("upload-initial", function() {
     }));
 });
 
-gulp.task("upload-widget", function() {
-    var indexPath = './sezzle.js'
-    gulp.src(indexPath)
-        .pipe(rename(buttonUploadName))
-        .pipe(s3({
-            Bucket: 'sezzle-shopify-application', //  Required
-            ACL: 'public-read'       //  Needs to be user-defined
-        }, {
-            // S3 Constructor Options, ie:
-            maxRetries: 5
-        }));
-});
+/**
+ * Tasks for the CSS
+ */
 
-gulp.task("post-button-to-widget-server", function() {
-    var options = {
-        method: 'POST',
-        uri: 'https://widget.sezzle.com/v1/javascript/price-widget/version',
-        body: {
-            'version_name': buttonUploadName
-        },
-        json: true
-    }
-    rp(options)
-        .then(function(body) {
-            console.log("Posted new version to shopify wrapper")
-        })
-        .catch(function(err) {
-            console.log("Post failed with shopify, ")
-            console.log(err);
-        })
-});
-
-gulp.task('post-button-css-to-wrapper', function() {
-	console.log("Posting css version to shopify gateway")
-    var options = {
-        method: 'POST',
-        uri: 'https://widget.sezzle.com/v1/css/price-widget/version',
-        body: {
-            'version_name': globalCssUploadName
-        },
-        json: true
-    }
-    rp(options)
-    .then(function(body) {
-        console.log("Posted new version to shopify wrapper")
-    })
-    .catch(function(err) {
-        console.log("Post failed with sezzle pay, ")
-        console.log(err);
-    })
+// cleans up dist directory
+gulp.task('cleancss', function() {
+  return del(['dist/global-css/**']);
 });
 
 // compiles scss and minifies
@@ -115,11 +63,89 @@ gulp.task('csscompile', function() {
     .pipe(gulp.dest('dist/global-css'))
 });
 
-// cleans up dist directory
-gulp.task('cleancss', function() {
-    return del(['dist/global-css/**']);
+gulp.task('cssupload', function() {
+  // bucket base url https://d3svog4tlx445w.cloudfront.net/
+  var indexPath = './dist/global-css/global.min.css'
+  gulp.src(indexPath)
+      .pipe(rename('shopify-app/assets/' + globalCssUploadName))
+      .pipe(s3({
+          Bucket: 'sezzlemedia', //  Required
+          ACL: 'public-read'       //  Needs to be user-defined
+      }, {
+          maxRetries: 5
+      }))
 });
 
-gulp.task('styles', ['cleancss', 'csscompile']);
-gulp.task("deploywidget", ["upload-widget", "post-button-to-widget-server"])
-gulp.task("deploycss", ["cssupload", "post-button-css-to-wrapper"])
+gulp.task('post-button-css-to-wrapper', function() {
+	console.log('Posting css version to shopify gateway')
+    var options = {
+        method: 'POST',
+        uri: 'https://widget.sezzle.com/v1/css/price-widget/version',
+        body: {
+            'version_name': globalCssUploadName
+        },
+        json: true
+    }
+    rp(options)
+    .then(function(body) {
+        console.log('Posted new version to shopify wrapper')
+    })
+    .catch(function(err) {
+        console.log('Post failed with sezzle pay, ')
+        console.log(err);
+    })
+});
+
+
+/**
+ * Tasks for the sezzle-js widget
+ */
+gulp.task('bundlejs', function() {
+  return gulp.src('src/sezzle.js')
+    .pipe(webpack({
+      output: {
+        filename: buttonUploadName,
+        libraryTarget: 'umd',
+        library: 'SezzleJS'
+      },
+      optimization:{
+        minimize: true, // <---- disables uglify.
+      }
+    }))
+    .pipe(gulp.dest('dist/'));
+});
+
+gulp.task('upload-widget', function() {
+  var indexPath = `./dist/${buttonUploadName}`
+  gulp.src(indexPath)
+      .pipe(s3({
+          Bucket: 'sezzle-shopify-application', //  Required
+          ACL: 'public-read'       //  Needs to be user-defined
+      }, {
+          // S3 Constructor Options, ie:
+          maxRetries: 5
+      }));
+});
+
+gulp.task('post-button-to-widget-server', function() {
+  var options = {
+      method: 'POST',
+      uri: 'https://widget.sezzle.com/v1/javascript/price-widget/version',
+      body: {
+          'version_name': buttonUploadName
+      },
+      json: true
+  }
+  rp(options)
+      .then(function(body) {
+          console.log('Posted new version to shopify wrapper')
+      })
+      .catch(function(err) {
+          console.log('Post failed with shopify, ')
+          console.log(err);
+      })
+});
+
+gulp.task('styles', gulp.series('cleancss', 'csscompile'));
+gulp.task('deploywidget', gulp.series('bundlejs', 'upload-widget', 'post-button-to-widget-server'))
+gulp.task('deploycss', gulp.series('cssupload', 'post-button-css-to-wrapper'))
